# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Discount System API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/discounts/store": {
            "post": {
                "tags": [
                    "Create Discounts"
                ],
                "summary": "Create Discount",
                "description": "Create a new discount",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "type",
                                    "value"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "WELCOME10"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "fixed or percentage"
                                    },
                                    "value": {
                                        "type": "string",
                                        "example": "10"
                                    },
                                    "min_cart_total": {
                                        "type": "string",
                                        "example": "50"
                                    },
                                    "applicable_products": {
                                        "type": "string",
                                        "example": [2, 6]
                                    },
                                    "applicable_categories": {
                                        "type": "string",
                                        "example": [3, 7]
                                    },
                                    "active_from": {
                                        "type": "string",
                                        "example": "2025-01-01 00:00:00"
                                    },
                                    "active_to": {
                                        "type": "string",
                                        "example": "2025-12-31 23:59:59"
                                    },
                                    "stackable": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Created Discounts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "example": "WELCOME10"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "fixed or percentage"
                                        },
                                        "value": {
                                            "type": "string",
                                            "example": "10"
                                        },
                                        "min_cart_total": {
                                            "type": "string",
                                            "example": "50"
                                        },
                                        "applicable_products": {
                                            "type": "string",
                                            "example": [2, 6]
                                        },
                                        "applicable_categories": {
                                            "type": "string",
                                            "example": [3, 7]
                                        },
                                        "max_uses": {
                                            "type": "integer",
                                            "example": "100"
                                        },
                                        "per_user_limit": {
                                            "type": "integer",
                                            "example": "3"
                                        },
                                        "first_time_only": {
                                            "type": "boolean"
                                        },
                                        "active_from": {
                                            "type": "string",
                                            "example": "2025-01-01 00:00:00"
                                        },
                                        "active_to": {
                                            "type": "string",
                                            "example": "2025-12-31 23:59:59"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not Found"
                    }
                }
            }
        },
         "/api/discounts/apply": {
            "post": {
                "tags": [
                    "Apply Discounts"
                ],
                "summary": "Apply Discounts on Cart",
                "description": "Apply Discounts on Cart",
                "operationId": "applyDiscount",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "cart"
                                ],
                                "properties": {
                                    "cart_total": {
                                        "type": "string",
                                        "example": 100
                                    },
                                    "cart_items": {
                                        "type": "string",
                                        "example": [
                                            { "product_id": 1, "category_id": 5, "price": 50 },
                                            { "product_id": 2, "category_id": 5, "price": 50 }
                                        ]
                                         
                                    },
                                    "discount_codes": {
                                        "type": "string",
                                        "example": ["WELCOME10"]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Discount Applied Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "discount_applied": {
                                            "type": "string",
                                            "example":10
                                        },
                                        "final_total": {
                                            "type": "string",
                                            "example": 90
                                        },
                                        "applied_discounts": {
                                            "type": "string",
                                            "example": ["WELCOME10"]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/discounts/": {
            "get": {
                "tags": [
                    "Discounts List"
                ],
                "summary": "Get list of discounts",
                "description": "Returns list of discounts",
                "operationId": "getDiscountsList",
                "responses": {
                    "200": {
                        "description": "Successfully returns List of all discounts",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Create Discounts",
            "description": "Create Discounts"
        },
        {
            "name": "Apply Discounts",
            "description": "Apply Discounts on Cart"
        },
        {
            "name": "Discounts List",
            "description": "Get Discounts List"
        }
    ]
}